+++ modules/build.js
@@ -14,8 +14,8 @@ module.exports = () => {
   console.info(' Building...');
 
   // Read all the data:
-  async.eachSeries(global.config.fileOrder, (file, callback) => {
-    fs.readFile(path.join(global.config.source, file), 'utf-8', (err, data) => {
+  async.eachSeries(global.config.buildOrder, (file, callback) => {
+    fs.readFile(path.join(path.dirname(global.config.source), file), 'utf-8', (err, data) => {
       if (err) return callback(err);
 
       allData[file] = data;
@@ -27,7 +27,7 @@ module.exports = () => {
     // Minify if necessary:
     minifyCode(allData, (data) => {
       let buildPath = global.config.output.path;
-      let buildName = global.config.output.name
+      let buildName = global.config.output.name;
       fs.writeFile(path.join(buildPath, buildName), data, 'utf-8', (err) => {
         if (err) {
           // If the dir does not exist make a new dir.
@@ -35,7 +35,7 @@ module.exports = () => {
             fs.mkdir(buildPath, (err) => {
               if (err) return console.erro(err);
               fs.writeFile(path.join(buildPath, buildName), data, 'utf-8', (err) => {
-                if (err) return console.erro(err);
+                if (err) return console.error(err);
               });
             });
           } else {
@@ -53,7 +53,7 @@ module.exports = () => {
         notify('Build Complete.', notifMessage);
         console.info('\n', timestamp, '- Build complete.');
         console.timeEnd(' Build Time');
-      })
-    })
+      });
+    });
   });
 }